requirements.txt
django>=5.0,<5.1
django-environ==0.11.2
django-paypal==2.1
